removed boxing
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/branches/cdk-1.2.x@13825 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/main/org/openscience/cdk/isomorphism/matchers/smarts/ExplicitConnectionAtom.java
@@ -58,8 +58,8 @@ public class ExplicitConnectionAtom extends SMARTSAtom {
 	 */
     public boolean matches(IAtom atom) {
         int ih = atom.getHydrogenCount() == CDKConstants.UNSET ? 0 : atom.getHydrogenCount();
-        int conn = ((Integer) atom.getProperty(CDKConstants.TOTAL_CONNECTIONS)).intValue() -
-                ((Integer) atom.getProperty(CDKConstants.TOTAL_H_COUNT)).intValue();
+        int conn = (Integer) atom.getProperty(CDKConstants.TOTAL_CONNECTIONS) -
+                (Integer) atom.getProperty(CDKConstants.TOTAL_H_COUNT);
         
         return numOfConnection == conn;
     }
